update to 123 ;aster repo from 2023/06/23

+++ vs/sdl2/build-dosbox.sh
@@ -1,4 +1,21 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+MAKE=""
+find_make()
+{
+  if test "$MAKE" = "" && \
+     command -v $1 >/dev/null && \
+     $1 --version | grep -q "GNU Make"; then
+    MAKE="$1"
+  fi
+}
+find_make make
+find_make gmake
+if test "$MAKE" = ""; then
+  echo "Couldn't find GNU Make!"
+  exit 1
+fi
+
 rm -Rfv linux-host || exit 1
 mkdir -p linux-host || exit 1
 
@@ -18,6 +35,9 @@ sys=`uname -s`
 
 if [ "$sys" == "Darwin" ]; then
 opts="--disable-video-x11"
+elif [ "$sys" != "Linux" ]; then
+# These are supported on BSDs but the SDL2 code assumes Linux
+opts="--disable-video-wayland --disable-libudev"
 fi
 if [ "$1" == "hx-dos" ]; then
 opts="--disable-video-opengl"
@@ -46,6 +66,6 @@ cat >>include/SDL_config.h <<_EOF
 _EOF
 fi
 
-make -j || exit 1
-make install || exit 1  # will install into ./linux-host
+$MAKE -j3 || exit 1
+$MAKE install || exit 1  # will install into ./linux-host
